src/requirements.txt
anyio==3.4.0
aprs3==8.0.0.post1
aprslib==0.7.0
asgiref==3.4.1
attrs==23.1.0
ax253==0.1.5.post1
bitarray==2.7.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
fastapi==0.70.1
gpxpy==1.5.0
h11==0.12.0
httptools==0.3.0
idna==3.3
importlib-metadata==6.5.0
Jinja2==3.0.3
kiss==7.0.0
kiss3==8.0.0
MarkupSafe==2.0.1
pydantic==1.8.2
pyserial==3.5
pyserial-asyncio==0.6
python-dotenv==0.19.2
PyYAML==6.0
redis==4.0.2
requests==2.26.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
zipp==3.15.0
